Refactored DTO folder structure and all file inputs related to DTOs
