+++ back_end/src/main/java/rikko/yugen/dto/artist/ArtistCreateDTO.java
@@ -1,4 +1,4 @@
-package rikko.yugen.dto;
+package rikko.yugen.dto.artist;
 
 import jakarta.validation.constraints.NotBlank;
