+++ back_end/src/main/java/rikko/yugen/dto/artist/ArtistDTO.java
@@ -1,5 +1,6 @@
-package rikko.yugen.dto;
+package rikko.yugen.dto.artist;
 
+import rikko.yugen.dto.user.UserDTO;
 import rikko.yugen.model.Artist;
 
 public class ArtistDTO {
