+++ back_end/src/main/java/rikko/yugen/dto/post/PostCreateDTO.java
@@ -1,4 +1,4 @@
-package rikko.yugen.dto;
+package rikko.yugen.dto.post;
 
 import java.util.Set;
