+++ back_end/src/main/java/rikko/yugen/dto/product/ProductCreateDTO.java
@@ -1,4 +1,4 @@
-package rikko.yugen.dto;
+package rikko.yugen.dto.product;
 
 import java.util.Set;
