+++ back_end/src/main/java/rikko/yugen/dto/product/ProductDTO.java
@@ -1,9 +1,12 @@
-package rikko.yugen.dto;
+package rikko.yugen.dto.product;
 
 import java.util.Collections;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import rikko.yugen.dto.artist.ArtistDTO;
+import rikko.yugen.dto.collection.CollectionDTO;
+import rikko.yugen.dto.series.SeriesDTO;
 import rikko.yugen.model.Product;
 
 public class ProductDTO {
