+++ back_end/src/main/java/rikko/yugen/dto/user/BaseUserDTO.java
@@ -1,0 +1,59 @@
+package rikko.yugen.dto.user;
+
+public class BaseUserDTO {
+    private Long id;
+    private String username;
+    private String displayName;
+    private String email;
+    private String image;
+    private Boolean isArtist;
+
+    // Getters and setters
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getDisplayName() {
+        return displayName;
+    }
+
+    public void setDisplayName(String displayName) {
+        this.displayName = displayName;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getImage() {
+        return image;
+    }
+
+    public void setImage(String image) {
+        this.image = image;
+    }
+
+    public Boolean getIsArtist() {
+        return isArtist;
+    }
+
+    public void setIsArtist(Boolean isArtist) {
+        this.isArtist = isArtist;
+    }
+}
