+++ back_end/src/main/java/rikko/yugen/dto/user/LoginResponseDTO.java
@@ -1,4 +1,4 @@
-package rikko.yugen.dto;
+package rikko.yugen.dto.user;
 
 import rikko.yugen.model.LoginResponse;
