+++ back_end/src/main/java/rikko/yugen/dto/user/UserLoginDTO.java
@@ -1,4 +1,4 @@
-package rikko.yugen.dto;
+package rikko.yugen.dto.user;
 
 public class UserLoginDTO {
     private String username;
